Rename utils common.
